fix: routing manager update (#2033)
| Q            | A
|--------------| ---
| Bug fix?     | yes
| New feature? | no
| BC-Break?    | no
| Backport     | 0.14
| License      | MIT

fix: routing manager update

+++ src/Enhavo/Bundle/RoutingBundle/Manager/RouteManager.php
@@ -35,17 +35,20 @@ class RouteManager
 
             $route = $resource->getRoute();
             $route->setContent($resource);
-        }
 
-        if (empty($route->getName())) {
-            $route->generateRouteName();
+            if (empty($route->getName())) {
+                $route->generateRouteName();
+            }
         }
 
         $this->autoGenerator->generate($resource);
 
-        if (empty($route->getStaticPrefix())) {
-            $resource->setRoute(null);
-            $this->em->remove($route);
+        if ($resource instanceof Routeable) {
+            $route = $resource->getRoute();
+            if ($route && empty($route->getStaticPrefix())) {
+                $resource->setRoute(null);
+                $this->em->remove($route);
+            }
         }
     }
 }
